Compiler for Ada in C
I built this Compiler during my compiler construction class in C. The
compiler makes sure the program is syntactically correct and then emits
register code to a text filed called machineCode.txt. The compiler was
built using lex and yacc.

It handles the following within the ada language: ( with patching )
- variables of integer and bool (local and non local)
- enums
- function calls with and without parameters ( local and non local)
- exceptions
- array types
- assignments
- expressions
- loops
- if/elsif/else